# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Connect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface RegisterRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  collegeId: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RefreshRequest {
  refreshToken: string;
}

export type Gender = (typeof Gender)[keyof typeof Gender];

export const Gender = {
  MALE: "MALE",
  FEMALE: "FEMALE",
  NON_BINARY: "NON_BINARY",
  PREFER_NOT_TO_SAY: "PREFER_NOT_TO_SAY",
} as const;

export interface College {
  id: string;
  name: string;
  domain: string;
  location: string;
}

export type VerificationStatus =
  (typeof VerificationStatus)[keyof typeof VerificationStatus];

export const VerificationStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  username?: string | null;
  email: string;
  bio?: string | null;
  gender?: Gender | null;
  age?: number | null;
  profilePhotos: string[];
  collegeId: string;
  college?: College | null;
  interests: string[];
  clubs: string[];
  isEmailVerified: boolean;
  verificationStatus: VerificationStatus;
  isProfilePublic: boolean;
  showAge: boolean;
  showGender: boolean;
  profileCompleted: boolean;
  isActive: boolean;
  lastActiveAt?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: UserProfile;
}

export interface PublicUserProfile {
  id: string;
  name: string;
  username?: string | null;
  bio?: string | null;
  gender?: Gender | null;
  age?: number | null;
  profilePhotos: string[];
  collegeId: string;
  college?: College | null;
  interests: string[];
  clubs: string[];
  isEmailVerified: boolean;
  profileCompleted: boolean;
}

export interface UpdateProfileRequest {
  name?: string;
  username?: string;
  bio?: string;
  gender?: Gender;
  age?: number;
  interests?: string[];
  clubs?: string[];
  profilePhotos?: string[];
  isProfilePublic?: boolean;
  showAge?: boolean;
  showGender?: boolean;
}

export interface UserStats {
  postsCount: number;
  matchesCount: number;
  likesReceived: number;
  commentsReceived: number;
}

export type PostType = (typeof PostType)[keyof typeof PostType];

export const PostType = {
  TEXT: "TEXT",
  IMAGE: "IMAGE",
  CONFESSION: "CONFESSION",
  EVENT: "EVENT",
} as const;

export interface Post {
  id: string;
  authorId?: string | null;
  author?: PublicUserProfile | null;
  collegeId: string;
  content?: string | null;
  mediaUrls: string[];
  postType: PostType;
  isAnonymous: boolean;
  likesCount: number;
  commentsCount: number;
  isLiked: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostRequest {
  content?: string;
  mediaUrls?: string[];
  postType: PostType;
  isAnonymous?: boolean;
}

export interface FeedResponse {
  posts: Post[];
  nextCursor?: string | null;
}

export interface LikeResponse {
  liked: boolean;
  likesCount: number;
}

export interface Comment {
  id: string;
  postId: string;
  authorId: string;
  author: PublicUserProfile;
  content: string;
  createdAt: string;
}

export interface AddCommentRequest {
  /** @minLength 1 */
  content: string;
}

export interface CommentsResponse {
  comments: Comment[];
  nextCursor?: string | null;
}

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus];

export const MessageStatus = {
  SENT: "SENT",
  DELIVERED: "DELIVERED",
  READ: "READ",
} as const;

export type MessageType = (typeof MessageType)[keyof typeof MessageType];

export const MessageType = {
  TEXT: "TEXT",
  IMAGE: "IMAGE",
} as const;

export interface Message {
  id: string;
  conversationId: string;
  senderId: string;
  receiverId: string;
  content?: string | null;
  mediaUrl?: string | null;
  messageType: MessageType;
  status: MessageStatus;
  createdAt: string;
}

export interface Conversation {
  id: string;
  otherUser: PublicUserProfile;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
}

export interface StartConversationRequest {
  userId: string;
}

export interface SendMessageRequest {
  content?: string;
  mediaUrl?: string;
  messageType: MessageType;
}

export interface MessagesResponse {
  messages: Message[];
  nextCursor?: string | null;
}

export interface RecommendedUser {
  user: PublicUserProfile;
  compatibilityScore: number;
  commonInterests: string[];
}

export type SwipeAction = (typeof SwipeAction)[keyof typeof SwipeAction];

export const SwipeAction = {
  LIKE: "LIKE",
  PASS: "PASS",
} as const;

export interface SwipeRequest {
  toUserId: string;
  action: SwipeAction;
}

export interface SwipeResponse {
  matched: boolean;
  matchId?: string | null;
}

export interface Match {
  id: string;
  otherUser: PublicUserProfile;
  compatibilityScore: number;
  matchedAt: string;
  conversationId?: string | null;
}

export interface MatchStats {
  totalMatches: number;
  pendingLikes: number;
  profileViews: number;
}

export type ReportReason = (typeof ReportReason)[keyof typeof ReportReason];

export const ReportReason = {
  SPAM: "SPAM",
  ABUSE: "ABUSE",
  HARASSMENT: "HARASSMENT",
  FAKE_PROFILE: "FAKE_PROFILE",
  NSFW: "NSFW",
  OTHER: "OTHER",
} as const;

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  PENDING: "PENDING",
  REVIEWED: "REVIEWED",
  ACTION_TAKEN: "ACTION_TAKEN",
} as const;

export interface ReportRequest {
  targetUserId?: string;
  targetPostId?: string;
  targetMessageId?: string;
  reason: ReportReason;
  description?: string;
}

export interface BlockRequest {
  userId: string;
}

export interface Report {
  id: string;
  reporterId: string;
  targetUserId?: string | null;
  targetPostId?: string | null;
  reason: ReportReason;
  description?: string | null;
  status: ReportStatus;
  createdAt: string;
}

export type AdminActionType =
  (typeof AdminActionType)[keyof typeof AdminActionType];

export const AdminActionType = {
  BAN_USER: "BAN_USER",
  DELETE_POST: "DELETE_POST",
  WARN_USER: "WARN_USER",
  DISMISS: "DISMISS",
} as const;

export interface AdminActionRequest {
  reportId: string;
  actionType: AdminActionType;
  notes?: string;
}

export interface AdminUserView {
  id: string;
  name: string;
  email: string;
  collegeId: string;
  college?: College | null;
  isActive: boolean;
  isBanned: boolean;
  reportCount: number;
  createdAt: string;
}

export type GetColleges200 = {
  colleges: College[];
};

export type GetFeedParams = {
  cursor?: string;
  postType?: PostType;
  limit?: number;
};

export type GetCommentsParams = {
  cursor?: string;
};

export type GetConversations200 = {
  conversations: Conversation[];
};

export type GetMessagesParams = {
  cursor?: string;
};

export type GetRecommendationsParams = {
  limit?: number;
};

export type GetRecommendations200 = {
  recommendations: RecommendedUser[];
};

export type GetMatches200 = {
  matches: Match[];
};

export type GetBlockedUsers200 = {
  blocked: PublicUserProfile[];
};

export type GetAdminReportsParams = {
  status?: ReportStatus;
};

export type GetAdminReports200 = {
  reports: Report[];
};

export type GetAdminUsersParams = {
  flagged?: boolean;
};

export type GetAdminUsers200 = {
  users: AdminUserView[];
};