# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Connect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- USERS ---
...
  /chat/conversations:
    get:
      operationId: getConversations
      tags: [chat]
      summary: Get all conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"
                required:
                  - conversations
  /chat/start:
    post:
      operationId: startConversation
      tags: [chat]
      summary: Start or get a conversation with a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationRequest"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # --- MATCH ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # --- AUTH SCHEMAS ---
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        collegeId:
          type: string
      required:
        - name
        - email
        - password
        - collegeId
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - accessToken
        - refreshToken
        - user

    # --- USER SCHEMAS ---
    Gender:
      type: string
      enum: [MALE, FEMALE, NON_BINARY, PREFER_NOT_TO_SAY]
    VerificationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
          nullable: true
        email:
          type: string
        bio:
          type: string
          nullable: true
        gender:
          $ref: "#/components/schemas/Gender"
          nullable: true
        age:
          type: integer
          nullable: true
        profilePhotos:
          type: array
          items:
            type: string
        collegeId:
          type: string
        college:
          $ref: "#/components/schemas/College"
          nullable: true
        interests:
          type: array
          items:
            type: string
        clubs:
          type: array
          items:
            type: string
        isEmailVerified:
          type: boolean
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        isProfilePublic:
          type: boolean
        showAge:
          type: boolean
        showGender:
          type: boolean
        profileCompleted:
          type: boolean
        isActive:
          type: boolean
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - profilePhotos
        - collegeId
        - interests
        - clubs
        - isEmailVerified
        - verificationStatus
        - isProfilePublic
        - showAge
        - showGender
        - profileCompleted
        - isActive
        - createdAt
    PublicUserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        gender:
          $ref: "#/components/schemas/Gender"
          nullable: true
        age:
          type: integer
          nullable: true
        profilePhotos:
          type: array
          items:
            type: string
        collegeId:
          type: string
        college:
          $ref: "#/components/schemas/College"
          nullable: true
        interests:
          type: array
          items:
            type: string
        clubs:
          type: array
          items:
            type: string
        isEmailVerified:
          type: boolean
        profileCompleted:
          type: boolean
      required:
        - id
        - name
        - profilePhotos
        - collegeId
        - interests
        - clubs
        - isEmailVerified
        - profileCompleted
...
    College:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        location:
          type: string
      required:
        - id
        - name
        - domain
        - location

    # --- POST SCHEMAS ---
...
    MessageStatus:
      type: string
      enum: [SENT, DELIVERED, READ]
    MessageType:
      type: string
      enum: [TEXT, IMAGE]
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        content:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
        messageType:
          $ref: "#/components/schemas/MessageType"
        status:
          $ref: "#/components/schemas/MessageStatus"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - senderId
        - receiverId
        - messageType
        - status
        - createdAt
    Conversation:
      type: object
      properties:
        id:
          type: string
        otherUser:
          $ref: "#/components/schemas/PublicUserProfile"
        lastMessage:
          type: string
          nullable: true
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - otherUser
        - unreadCount
        - createdAt
    StartConversationRequest:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
        mediaUrl:
          type: string
        messageType:
          $ref: "#/components/schemas/MessageType"
      required:
        - messageType
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: string
          nullable: true
      required:
        - messages

    # --- MATCH SCHEMAS ---
...