# lib/api-zod/src/generated/types/college.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Connect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface College {
  id: string;
  name: string;
  domain: string;
  location: string;
}